+++ @types/bin/bundle-scss.d.ts
@@ -1,0 +1,3 @@
+declare const app: any;
+declare const cfg: any;
+declare const bundleScss: any;