+++ @types/helpers/constants.d.ts
@@ -1,0 +1,5 @@
+export declare const config: {
+    defaultPriority: string[];
+    sassImportRegex: RegExp;
+    fileType: string;
+};