+++ @types/index.d.ts
@@ -1,0 +1,2 @@
+declare const _default: (mask: string | string[], dest: string, sort?: string[]) => Promise<string | void>;
+export = _default;