+++ src/index.ts
@@ -2,13 +2,15 @@ import { writeFile, statSync, readFileSync } from 'fs';
 import { resolve, join, dirname } from 'path';
 import * as globby from 'globby';
 import { StringDecoder } from 'string_decoder';
+import { config } from './helpers/constants';
+import { Sort } from './helpers/Sort';
 
 const decoder = new StringDecoder('utf8');
 const log = (info: Error | string) => console.log(info);
 const isFile = (f: string) => statSync(f).isFile();
 
 const removeImports = (content: string) =>
-  content.replace(/@import ['"]([^'"]+)['"];/g, '');
+  content.replace(config.sassImportRegex, '');
 
 const readSync = (filePath: string) => readFileSync(filePath, 'utf8');
 const getUniqueScss = (files: Array<string>) => {
@@ -18,20 +20,23 @@ const getUniqueScss = (files: Array<string>) => {
       return getImports(readSync(file), baseDir);
     })
     .reduce((acc, curr) => acc.concat(curr), []);
-  const uniqueArr = [...new Set(scssImports.map(el => el.path))];
+  const allImports = [...scssImports, ...files];
 
-  return [...uniqueArr, ...files];
+  return [...new Set(allImports)];
 };
 
-const getImports = (content: string, baseDir: string) => {
-  const regex = /@import ['"]([^'"]+)['"];/g;
-  let imports = [];
+const getImports = (
+  content: string,
+  baseDir: string,
+  imports: Array<string> = []
+) => {
   let match;
-  while ((match = regex.exec(content)) !== null) {
-    imports.push({
-      path: defineExtension(join(baseDir, match[1])),
-      stringToReplace: match[0],
-    });
+  while ((match = config.sassImportRegex.exec(content)) !== null) {
+    const pathFile = defineExtension(join(baseDir, match[1]));
+    if (!imports.some(el => el === pathFile)) {
+      imports.push(pathFile);
+      getImports(readSync(pathFile), dirname(pathFile), imports);
+    }
   }
 
   return imports;
@@ -55,20 +60,27 @@ const defineExtension = (filePath: string) => {
   if (isFile(justScss)) {
     return justScss;
   }
-  console.error(`⛔ ⛔ ⛔ No file for module ${filePath}`);
+  throw new Error(`⛔ ⛔ ⛔ No file for module ${filePath}`);
 };
 
-export = (mask: string[] | string, dest: string) => {
+export = (
+  mask: string[] | string,
+  dest: string,
+  sort: string[] = config.defaultPriority
+) => {
   const fullPath = join(process.cwd());
+  const sortOrder = Array.isArray(sort) ? sort : [sort];
+  const sortInstance = new Sort(sortOrder);
   if (!mask || !mask.length) {
-    console.error('⛔ ⛔ ⛔ Please provide the src for concat method');
+    throw new Error('⛔ ⛔ ⛔ Please provide the src for concat method');
   }
   const searchMask = Array.isArray(mask) ? mask : [mask];
   return globby(searchMask).then(paths => {
     const files = paths.map(file => join(fullPath, file));
 
     const unique = getUniqueScss(files);
-    const buffers = unique.map(file => {
+    const sorted = sortInstance.sort(unique);
+    const buffers = sorted.map(file => {
       return readFileSync(file);
     });
     const buff = Buffer.concat(buffers);
@@ -78,7 +90,7 @@ export = (mask: string[] | string, dest: string) => {
       log(`⏳ ⏳ ⏳ Saving result to ${dest}...`);
       const utf = removeImports(utfFormat);
       return writeAsync(dest, utf)
-        .then(res => {
+        .then(() => {
           log(`🚀 🚀 🚀 SAVED SUCCESSFULLY \nPlease check ${dest}`);
           return utf;
         })
@@ -87,5 +99,6 @@ export = (mask: string[] | string, dest: string) => {
         });
     }
     log('📁 Please provide destination option ');
+    throw new Error('📁 Please provide destination option ');
   });
 };