+++ test/test.js
@@ -1,11 +1,11 @@
-const bundleScss = require('../index');
+const bundleScss = require('../dist/index');
 const assert = require('assert');
 const read = require('fs').readFileSync;
 const rimraf = require('rimraf');
 
-describe('Concatenate', () => {
+describe('bundle-scss', () => {
   rimraf.sync('./test/result.scss');
-  it('scss into one and compare', (done) => {
+  it('Should concatenate scss into one and compare', (done) => {
     bundleScss(["./**/*.theme.scss"], './test/result.scss').then(f => {
       assert.equal(f, read('./test/compare.scss', 'utf8'));
       done()